+++ rent-a-car/src/main/java/com/tobeto/rentacar/services/concretes/PaymentManager.java
@@ -14,6 +14,7 @@ import java.util.List;
 @Service
 public class PaymentManager implements PaymentService {
     private final PaymentRepository paymentRepository;
+    private final String errorMessage = "Invalid information";
 
     public PaymentManager(PaymentRepository paymentRepository) {
         this.paymentRepository = paymentRepository;
@@ -21,6 +22,9 @@ public class PaymentManager implements PaymentService {
 
     @Override
     public void add(AddPaymentRequest addPaymentRequest){
+        if(paymentRepository.existsByPaymentTypeStartingWith(addPaymentRequest.getPaymentType())){
+            throw new RuntimeException(errorMessage);
+        }
         Payment payment = new Payment();
         payment.setPaymentType(addPaymentRequest.getPaymentType());
         paymentRepository.save(payment);
@@ -28,6 +32,9 @@ public class PaymentManager implements PaymentService {
 
     @Override
     public void update(UpdatePaymentRequest updatePaymentRequest){
+        if(!paymentRepository.existsById(updatePaymentRequest.getId())){
+            throw new RuntimeException("Invalid id");
+        }
         Payment paymentToUpdate = paymentRepository.findById(updatePaymentRequest.getId()).orElseThrow();
         paymentToUpdate.setPaymentType((updatePaymentRequest.getPaymentType()));
         paymentRepository.save(paymentToUpdate);
@@ -35,6 +42,9 @@ public class PaymentManager implements PaymentService {
 
     @Override
     public void delete(int id){
+        if(!paymentRepository.existsById(id)){
+            throw new RuntimeException(errorMessage);
+        }
         Payment paymentToDelete = paymentRepository.findById(id).orElseThrow();
         paymentRepository.delete(paymentToDelete);
     }
